WIP: cmTarget: ban INTERFACE_SOURCES from objlibs
INTERFACE_SOURCES is a special beast anyways in that the following setup
is likely to cause duplicate symbols:

    add_library(a INTERFACE)
    target_sources(a INTERFACE sym1.c)
    target_include_directories(a INTERFACE some/path/for/sym1)
    target_compile_definitions(a INTERFACE CLIENT_DEFINE)

    add_library(b SHARED shared.c)
    target_link_libraries(b PUBLIC a) # Needs CLIENT_DEFINE

    add_executable(exe main.c)
    target_link_libraries(exe b)

Here, sym1.c is compiled as part of 'b' and 'exe'. Issues arise when a
either has global statics within the source file (ld-linux.so will
construct it once, but try to destroy it twice), exports symbols
(multiple definitions), etc. This is especially prone to abuse with
OBJECT libraries, so block its usage.

The real fix is to make INTERFACE_SOURCES only applicable when directly
linked and not transitive, but that is a bigger change.